# openapi.yaml
openapi: 3.1.0
info:
  title: BugTraceAI CLI API
  description: |
    REST API for the BugTraceAI autonomous security scanning platform.

    BugTraceAI CLI is a headless, AI-powered vulnerability scanner that discovers and validates
    security issues using specialized agents (XSS, SQLi, SSRF, IDOR, LFI, RCE, XXE, JWT, and more).
    This API provides full scan lifecycle management, real-time WebSocket streaming, report generation,
    and runtime configuration.

    **Local beta deployment only.** No authentication is required because access is assumed to be
    limited to the local machine or a trusted LAN. Do not expose the API directly to the Internet.
    ModelLab uses the configured OpenRouter API key and may incur provider costs.
    WebSocket endpoints are available for real-time scan events, findings, and ModelLab progress.

    **Note:** The following endpoints are documented in code but not yet in this spec:
    - POST /api/auth/validate (YAML auth configuration validation)
    - PATCH /api/scans/{scan_id}/auth-config (Update scan authentication)
    - POST /api/tools/totp-verify (TOTP/2FA verification)
  version: 4.0.0-rc1
  contact:
    name: BugTraceAI
    url: https://bugtraceai.com
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8000
    description: Local development
  - url: http://{host}:8000
    description: Custom deployment
    variables:
      host:
...
paths:
...
  /api/scans/{scan_id}/report/{format}:
    get:
      tags: [reports]
      summary: Download report
      description: Download the scan report in the specified format.
      operationId: downloadReport
      parameters:
        - $ref: "#/components/parameters/ScanId"
        - name: format
          in: path
          required: true
          schema:
            type: string
            enum: [html, json, markdown]
          description: Report output format
      responses:
        "200":
          description: Report file
          content:
            text/html:
              schema:
                type: string
            application/json:
              schema:
                type: object
            text/markdown:
              schema:
                type: string
          headers:
            Content-Disposition:
              schema:
                type: string
              description: "attachment; filename=bugtrace_report_{scan_id}.{ext}"
        "400":
          description: Invalid format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Scan or report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/scans/{scan_id}/files/{filename}:
    get:
      tags: [reports]
      summary: Get report file
      description: |
        Serve individual files from a scan's report directory.
        Path traversal is prevented — the file must be within the report directory.
      operationId: getReportFile
      parameters:
        - $ref: "#/components/parameters/ScanId"
        - name: filename
          in: path
          required: true
          schema:
            type: string
          description: "Filename within the report directory (e.g. final_report.md)"
      responses:
        "200":
          description: File contents with auto-detected content type
        "400":
          description: Invalid filename (path traversal attempt)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Report directory or file not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    ScanId:
      name: scan_id
      in: path
      required: true
      schema:
        type: integer
      description: Unique scan identifier
  schemas:
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Human-readable error message
        error_code:
          type: string
          description: Machine-readable error code
...